+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Any non-code changes should be prefixed with `(docs)`.
 See `PUBLISH.md` for instructions on how to publish a new version.
 -->
 
+- (minor) Use shared toolbar for code block label + buttons
 
 ## v1.13.0 - 6056063